Report on light detector using NAND gate
AIM of the project:- To detect the light when the circuit is exposed to brightness.
In this project, we have designed a light detector by using NAND gates. Light detectors are the circuits used to detect the presence of light, when a bright light falls on the circuit. Eg:- a flash light, the LED turns ON.
A photoresistor is the major component used for the detection of light. Since, it is altered in the proportion equal to the amount of the light exposed to. In case of drakness, there is high resistance but in case of bright light, the resistance decreases. This is the working principle of light detector.
If there is a voltage divider circuit along with the fixed resistor, we will be able to exploit the altering resistance and its behaviour. This means, if one of the inputs is zero then the output is HIGH which means the output is taken by Vcc and power is applied to load. If both the inputs are HIGH, then the output pf NAND gate is LOW which means the output will be drawn to the ground terminal and load will not be powered.
OUTCOME:-When the LDR is exposed with the light the resistance of LDR decreases, as a result the led glows.