Report on Car number plate detection using MATLAB
In this project we have detected and identified vehicle number that will help  in the detection of car number plates easily.
This mainly follows Sobel edge detection method and morphlogical operation for better approach, and it is simplified to use all the numbers and letters which are in segmented fashion as it apperas in vehicle number plate. Numbers and characters are reconginsed using neutral netsworks and bounding box method. Our main intention is to locate the region of car number plate in proper way and add to segment all the numbers and letters to identify each number singly. Systems like ANPR, OCR,CRS are excited in this.
The image of the car number plate is captured by the camera of ANPR system and then processed through multiple number of algorithms and then finally we get outputs of the detected number plate. These are widely used in highways and airports etc.