Report on Smart street light system using IR sensor
The project on "Automactic street light system"  is being reviewed.
In this project light automatically turn ON when necessary and turn OFF themselves when not needed. presently, in the entire world, a huge amount of energy is consumed by normal street lamps which do not automatically turn OFF when its bright. 
Our project "Automatic smart street light system" consists of LED lights, IR sensors, and also a short distance communication network. Before the pedestrians and vehicles come, the lights turn ON and then automatically turn OFF once they cross. By doing this power is reduced.
It is basically an intelligent way of controlling the street lights to optimise the power consumption issues late in night. Hence, for this sake the innovation in automatic street lighting on comparision with the conventional street lighting has proven to me more effective and reliable as advised by the government ofindia as well.