Report on Fire Fighting Robot using Arduino UNO
Increasing population and technological growth and development all across the world has led to various problems such as accidents. In which fire accidents hold the major part.
● Till date fire accidents is a hazardous act that led to numerous major consequences. We rely on human resources till now. They are fire fighters. They risk their lives to 
extinguish it. Detecting fire and extinguishing it is a dangerous job that puts the life of a fire fighter at risk. There are many fire accidents which fire fighter had to lose their 
lives
● Fire accidents are one such major problems. To deal with this robotic industry has emerged with the project that can replace human resources with the machine called robot.Fire fighting robots can be used.
● Detecting fire and extinguishing it has been a difficult task. In our project a fire extinguishing robot has been proposed which detects the fire location , navigates 
through it and extinguishes the fire by pumping water.
● In this project we are implementing a robot that detects the presence of fire in 3 directions i.e. straight, left and right. The robot is controlled automatically without 
using any remote controller for its control.